% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twitter_networks.R
\name{create_community_table}
\alias{create_community_table}
\title{Summary about clusters
.}
\usage{
create_community_table(network, communities)
}
\arguments{
\item{network}{igraph object}

\item{communities}{communities of network}
}
\value{
Data frame with summary
}
\description{
This function creates summary every cluster in network
}
