% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twitter_networks.R
\name{create_network}
\alias{create_network}
\title{Create PM's Network}
\usage{
create_network(
  partia,
  poslowie,
  tweets,
  whole_parliament = F,
  include_retweets = T,
  include_replies = T,
  include_quotes = T,
  only_subparty = F
)
}
\arguments{
\item{partia}{Name of party (parliament club) which network will be made}

\item{poslowie}{Dataframe with all PM's twitters and party/subparty affiliations}

\item{tweets}{Dataframe with tweets data}

\item{whole_parliament}{Logical indicator if function will create whole parliament network, default set to False}

\item{include_retweets}{Logical indicator if function will include retweets from tweets file, default set to True}

\item{include_replies}{Logical indicator if function will include replies from tweets file, default set to True}

\item{include_quotes}{Logical indicator if function will include quotes from tweets file, default set to True}

\item{only_subparty}{Logical indicator if partia argument is name of parliament club or party, default set to False}
}
\value{
A igraph object representing network
}
\description{
This function creates ipgraph object based on interactions (retweets, replies, quotes)
in tweets file.
}
