% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twitter_networks.R
\name{make_undirected}
\alias{make_undirected}
\title{Convert network to undirected}
\usage{
make_undirected(network)
}
\arguments{
\item{network}{igraph object to convert}
}
\value{
A igraph object representing undirected network
}
\description{
This function makes given network undirected, simplifies it and removes unconnected verticies
}
