% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twitter_networks.R
\name{summary_interactions_among_clusters}
\alias{summary_interactions_among_clusters}
\title{Summary of inter cluster connections}
\usage{
summary_interactions_among_clusters(network, communities)
}
\arguments{
\item{network}{igraph object}

\item{communities}{communities of network}
}
\value{
Data frame with summary
}
\description{
This function creates summary for every verticle in network
showing sum of edges weights for every network .
}
