% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twitter_networks.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize network
.}
\usage{
visualize_network(network)
}
\arguments{
\item{network}{igraph object}
}
\value{
visualization
}
\description{
This function will visualize given network using visnetwork package
}
