% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twitter_networks.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize network
.}
\usage{
visualize_network(network, layout = "layout_nicely", show_legend = T)
}
\arguments{
\item{network}{igraph object}

\item{layout}{igraph layout algorithm default: "layout_nicely"}

\item{show_legend}{logical indicator if function will show legend for graph verticies colors}
}
\value{
visualization
}
\description{
This function will visualize given network using visnetwork package
}
